--- src/main/java/ToDo.java
@@ -1,16 +1,0 @@
-public class ToDo extends Task {
-    public ToDo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String storedString() {
-        return "T | " + super.storedString();
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
-
