+++ src/main/java/duke/Storage.java
@@ -1,4 +1,10 @@
-import exceptions.DukeInvalidFileException;
+package duke;
+
+import duke.exceptions.DukeInvalidFileException;
+import duke.tasks.Deadline;
+import duke.tasks.Event;
+import duke.tasks.Task;
+import duke.tasks.ToDo;
 
 import java.io.File;
 import java.io.FileNotFoundException;
