+++ src/main/java/duke/exceptions/DukeInvalidFileException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package duke.exceptions;
 
 public class DukeInvalidFileException extends DukeException{
     public DukeInvalidFileException() {
