+++ src/main/java/duke/exceptions/DukeInvalidInputException.java
@@ -1,0 +1,7 @@
+package duke.exceptions;
+
+public class DukeInvalidInputException extends DukeException {
+    public DukeInvalidInputException() {
+        super("This is nonsense.");
+    }
+}
