+++ src/main/java/duke/exceptions/DukeInvalidTimeException.java
@@ -1,0 +1,7 @@
+package duke.exceptions;
+
+public class DukeInvalidTimeException extends DukeException{
+    public DukeInvalidTimeException() {
+        super("As much as it would be nice to live free from the shackles of time... you don't.");
+    }
+}
