+++ src/main/java/duke/tasks/Task.java
@@ -1,0 +1,37 @@
+package duke.tasks;
+
+public class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "X" : " "); // mark done task with X
+    }
+
+    public String getStatusBinary() {
+        return (isDone ? "1" : "0");
+    }
+
+    public void markDone() {
+        this.isDone = true;
+    }
+
+    public void markUndone() {
+        this.isDone = false;
+    }
+
+    public String storedString() {
+        return this.getStatusBinary() + " | " + this.description;
+    }
+
+    @Override
+    public String toString() {
+        return "[" + this.getStatusIcon() + "]" + " " + this.description;
+    }
+}
+
