+++ src/main/java/duke/tasks/ToDo.java
@@ -1,0 +1,18 @@
+package duke.tasks;
+
+public class ToDo extends Task {
+    public ToDo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String storedString() {
+        return "T | " + super.storedString();
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+}
+
