--- src/main/java/exceptions/DukeException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class DukeException extends Exception {
-    public DukeException(String msg) {
-        super(msg);
-    }
-}
