--- src/main/java/exceptions/DukeInvalidIndexException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class DukeInvalidIndexException extends DukeException{
-    public DukeInvalidIndexException() {
-        super("You don't have a task at that index. What are you trying to do?");
-    }
-}
