--- src/main/java/exceptions/DukeInvalidInputException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class DukeInvalidInputException extends DukeException {
-    public DukeInvalidInputException() {
-        super("This is nonsense.");
-    }
-}
