--- src/main/java/exceptions/DukeInvalidTimeException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class DukeInvalidTimeException extends DukeException{
-    public DukeInvalidTimeException() {
-        super("As much as it would be nice to live free from the shackles of time... you don't.");
-    }
-}
